requirements.txt
pywabot
google-genai
python-dotenv
